rename network -> remote
