+++ app/src/main/java/io/minoro75/heremapsweatherapp/remote/RemoteDataSource.kt
@@ -1,4 +1,4 @@
-package io.minoro75.heremapsweatherapp.network
+package io.minoro75.heremapsweatherapp.remote
 
 import io.minoro75.heremapsweatherapp.domain.Weather
 import kotlinx.coroutines.flow.Flow
